# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /run:
    post:
      operationId: runCode
      tags: [execution]
      summary: Execute code (buffered JSON response)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunRequest"
      responses:
        "200":
          description: Execution result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /run/stream:
    post:
      operationId: runCodeStream
      tags: [execution]
      summary: Execute code with real-time SSE streaming output
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunRequest"
      responses:
        "200":
          description: Server-Sent Events stream of execution output
          content:
            text/event-stream:
              schema:
                type: string
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List user's saved projects
      parameters:
        - name: x-user-key
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectListResponse"
        "401":
          description: Missing user key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      operationId: createProject
      tags: [projects]
      summary: Save a new project
      parameters:
        - name: x-user-key
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: x-user-key
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateProject
      tags: [projects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: x-user-key
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
      responses:
        "200":
          description: Project updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteProject
      tags: [projects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: x-user-key
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  id:
                    type: string
                required: [success, id]
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    RunRequest:
      type: object
      properties:
        language:
          type: string
        code:
          type: string
        filename:
          type: string
      required: [language, code]
    RunResult:
      type: object
      properties:
        stdout:
          type: string
        stderr:
          type: string
        exitCode:
          type: integer
        duration:
          type: integer
        error:
          type: string
          nullable: true
        html:
          type: string
          nullable: true
      required: [stdout, stderr, exitCode, duration]
...
    CreateProjectRequest:
      type: object
      properties:
        name:
          type: string
        projectType:
          type: string
        files:
          type: object
          additionalProperties:
            type: string
      required: [name, files]
    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
        projectType:
          type: string
        files:
          type: object
          additionalProperties:
            type: string
    Project:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        projectType:
          type: string
        files:
          type: object
          additionalProperties:
            type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, projectType, createdAt, updatedAt]
    ProjectSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        projectType:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, projectType, createdAt, updatedAt]
    ProjectListResponse:
      type: object
      properties:
        projects:
          type: array
          items:
            $ref: "#/components/schemas/ProjectSummary"
      required: [projects]
    ProjectResponse:
      type: object
      properties:
        project:
          $ref: "#/components/schemas/Project"
      required: [project]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]