# lib/api-zod/src/generated/types/projectListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
import type { ProjectSummary } from "./projectSummary";

export interface ProjectListResponse {
  projects: ProjectSummary[];
}