# lib/api-zod/src/generated/types/projectResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
import type { Project } from "./project";

export interface ProjectResponse {
  project: Project;
}