# lib/api-zod/src/generated/types/projectSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */

export interface ProjectSummary {
  id: string;
  name: string;
  projectType: string;
  createdAt: string;
  updatedAt: string;
}