# lib/api-zod/src/generated/types/runResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */

export interface RunResult {
  stdout: string;
  stderr: string;
  exitCode: number;
  duration: number;
  error?: string | null;
  html?: string | null;
}