# lib/api-zod/src/generated/types/updateProjectRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
import type { UpdateProjectRequestFiles } from "./updateProjectRequestFiles";

export interface UpdateProjectRequest {
  name?: string;
  projectType?: string;
  files?: UpdateProjectRequestFiles;
}